+++ src/data-processor/calc-positions.ts
@@ -103,7 +103,7 @@ class CalcPositions {
         const data = await resources.texture.read(0, 0, resources.texture.width, resources.texture.height, {
             renderTarget: resources.renderTarget,
             data: resources.data,
-            immediate: false
+            immediate: this.device.isWebGPU
         });
 
         return data as Float32Array;