+++ src/data-processor/select-by-range.ts
@@ -193,7 +193,7 @@ class SelectByRange {
         const data = await resources.texture.read(0, 0, resources.texture.width, resources.texture.height, {
             renderTarget: resources.renderTarget,
             data: buffer,
-            immediate: false
+            immediate: this.device.isWebGPU
         });
 
         return data as Uint8Array;